requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
langchain-google-genai
langchain-anthropic
langchain_experimental
langchain_cohere
openai
anthropic
cohere
streamlit
python-dotenv
matplotlib
numpy
pandas
seaborn
tabulate
openpyxl
